backend/contacts/serializers.py
from rest_framework import serializers
from contacts.models import Contact

# Contact Serializer
class ContactSerializer(serializers.ModelSerializer):
  class Meta:
    model = Contact
    fields = '__all__'