backend/requirements.txt
asgiref==3.2.10
autopep8==1.5.3
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.7
django-cors-headers==3.4.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
djangorestframework-api-key==2.0.0
idna==2.10
Pillow==7.2.0
psycopg2==2.8.5
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
twilio==6.45.1
urllib3==1.25.10
whitenoise==5.2.0
